first reduction
