+++ src/lrg/GeneticAlgorithms/BasicGA.jl
@@ -7,14 +7,7 @@ genetic mutation and recombination).
 
 Authors: Alina Arneth, Michael Staab, Benedikt Traut, Adrian Wild 2022.
 """
-
-include("../utils/Mutations.jl")
-# include("../utils/Selections.jl")
-include("../utils/TestSelections.jl") # TODO: change import to real Selections module
-include("../utils/Recombinations.jl")
-include("../utils/Objectives.jl")
-
-using .Mutations, .Recombinations, .Selections, .Objectives
+using Statistics
 using Agents, Random
 using InteractiveDynamics, GLMakie
 
@@ -45,21 +38,13 @@ end
 
 Initializes the simulation.
 """
-function initialize(; N=100, M=1, mut::Mutation=DummyMutation(), obj::Objective=dummyObjective(),
-	recomb::Recombination=DummyRecombination(), sel::Selection=DummySelection(), seed=42, genome_length=128)
+function initialize(; N=100, M=1, seed=42, genome_length=128)
 
 	Random.seed!(seed);
 	space = GridSpace((M, M); periodic=false)
-	properties = Dict(
-		:mutations => mut,
-		:recombinations => recomb,
-		:selection => sel,
-		:objective => obj
-	)
 
 	model = ABM(
 		BasicGAAgent, space;
-		properties
 	)
 	for n in 1:N
         agent = BasicGAAgent(n, (1, 1), BasicGAAlleles.(rand([0,1], genome_length)))
@@ -79,15 +64,13 @@ function model_step!(model)
 	population = map(agent -> Int8.(agent.genome), allagents(model))
 	nAlleles = length(random_agent(model).genome)
 	# get fitness matrix:
-	# popFitness = map(genome -> fitness(genome), population)
-	popFitness = map(genome -> max(sum(genome), 100), population)
+	popFitness = fitness(population)
 	# Selection:
-	# selectionWinners = performSelection(model.selection, popFitness)
-	performSelection(model.selection)
+	selectionWinners = population[performSelection(popFitness), 1:end]
 	# Recombination:
 	# popₙ = recombine(model.recombination, selectionWinners)
 	# Mutation:
-	# popₙ = mutate(model.mutation, popₙ)
+	popₙ = mutate(popₙ, model.mu, model.obj)
 	# TODO: remove mock code that just makes all Alleles to ones and replace with actual mutation
 	for agent ∈ allagents(model)
 		agent.genome = fill(BasicGAAlleles(1), nAlleles)
@@ -105,4 +88,131 @@ function simulate()
 	model = initialize()
 	data, _ = run!(model, dummystep, model_step!, 5; adata=[a -> sum(Int.(a.genome))])
 	return data
+end
+
+"""
+	mutate!(genpool::Matrix{T}, alleles::Vector{T}, mutation::BitFlip) where {T <: Integer}
+
+Mutates the given genpool with Integer elements (alleles) according to the BitFlip mutation (see above), 
+but with the subsequent changes:
+	1. For each locus, a corresponding allele probability is fetched
+	2. Each allele probability is converted into a allele index
+	3. For mutation, each orginal allele at a selected locus is replaced by the allele 
+	   at the corresponding index
+"""
+function mutate!(genpool::Matrix{T}, alleles::Vector{T}; mu, casino) where {T <: Integer}
+	# Only expend effort on mutating if it is really wanted:
+	if mu > 0
+		# Get population size and number of genes per individual:
+		nIndividuals, nGenes = size(genpool)
+
+		# Find loci to mutate:
+		loci = draw(casino, nIndividuals, nGenes, mu)
+
+		# Fetch a allele probability to generate allele indices for locusmutation:
+		alleleIdx = Int64.(ceil.(
+			draw(casino, nIndividuals, nGenes) .* length(alleles)
+		))
+
+		# Now mutate the population:
+		for i ∈ 1:nIndividuals
+			# Modifiy the selected loci:
+			genpool[i, loci[i, :]] = alleles[alleleIdx[i, loci[i, :]]];
+		end
+	end
+
+	return genpool
+end
+
+#---------------------------------------------------------------------------------------------------
+
+"""
+	mutate!(genpool::Matrix{T}, mutation::BitFlip) where {T <: Enum}
+
+Mutates the given genpool of alleles (represented by enums) by modifying the underlying integers 
+(see above: mutate!(genpool::Matrix{T}, alleles::Vector{T}, mutation::BitFlip) where {T <: Integer})
+"""
+function mutate!(genpool::Matrix{T}, mu, casino) where {T <: Enum}
+	# Determine possible alleles
+	alleles = eltype(genpool);
+
+	# Get alleles and convert them to integers:
+	intAlleles = [Int(i) for i in instances(alleles)]
+	
+	# Perform mutation on the genpool of the population:
+	mutatedGenpool = mutate!(Int.(genpool), intAlleles, mu, casino)
+
+	# Convert integers back to alleles and modfiy original matrix:
+	return genpool .= alleles.(mutatedGenpool)
+end
+
+# -----------------------------------------------------------------------------------------
+"""
+	mepi(x)
+
+Watson's maximally epistatic objective function.
+"""
+function mepi(genome::BitVector)
+	dim = length(genome)
+
+	if dim == 1
+		1
+	else
+		# mepi is minimized if allels are the same
+		penality = all(genome) || !any(genome) ? 0 : 1
+		# Form product of the first and second halves of genome separately:
+		halflen = div(dim, 2)
+		dim * penality + mepi(genome[1:halflen]) + mepi(genome[halflen + 1:end])
+	end
+end
+
+"""
+	fitness(sga)
+
+Calculate normalised fitness based on the Objective function. fitness is a column vector of normalised 
+fitnesses of sga population, minus all sub-sigma-scaled individuals (see Mitchell p.168).
+Negative sigma-scaling maximises the objective function; higher magnitudes raise the fitness pressure. 
+evaluations is a colum vector of evaluations of the population. 
+"""
+function fitness(genpool::BitMatrix) 
+
+	npop = size(genpool)[1]
+
+	# Calculate the current objective evaluations of the population:
+	evaluations = mepi.([genpool[i,:] for i = 1:npop]) # = mepi.(genpool) when genpool::Vector{BitVector}
+	# Normalise the evaluations into frequencies:
+	sigma = std(evaluations)				# Standard deviation
+	if sigma == 0
+		# Singular case: all evaluations were equal to the mean:
+		fitness = ones(npop)
+	else
+		# Exorcise all evaluations worse than one standard
+		# deviations above mean value:
+		fitness = 1 .+ (mean(evaluations) .- evaluations) ./ (sigma)
+		fitness[fitness .<= 0] .= 0
+	end
+	
+	(fitness,evaluations)					# Return value
+end
+
+"""
+    This implementation of tournamentSelection expects a matrix of fitness values with the datatype float64
+"""
+function performSelection(popFitness::Array)
+    nPop = length(popFitness)
+    parents = Array(2*nPop)
+
+    for i in 1:2*nPop
+        
+        "Could be solved with sample function from StatsBase.jl"
+        firstFigther = rand(1:nPop)
+        secondFighter = rand(1:nPop)
+        
+        if popFitness[firstFigher] > popFitness[secondFigher]
+            parents[i] = firstFighter
+        else
+            parents[i] = secondFighter
+        end
+    end
+    return parents
 end
